feat(save): save boolean value to know if a map script has been executed or not

+++ src/game.h
@@ -4,8 +4,6 @@
 #ifndef STUGE_GAME_H
 #define STUGE_GAME_H
 
-#include <stdint.h>
-
 #include <lua.h>
 #include <raylib.h>
 
@@ -65,7 +63,7 @@ typedef struct {
     GameState state;
     GameState prev_state;
 
-    size_t map_idx;
+    int map_idx;
     Map* map;
     tmx_resource_manager* tmx_resman;
